+++ CHANGES.md
@@ -6,6 +6,7 @@
 - When calling `set-default`, you can now pass `--allow-undefined` to set the
   default to a version that doesn't exist yet
 - Add global-level `-q` / `--quiet` option to suppress warning messages
+- Add support for handling `!relative` in `mkdocs.yml`
 
 ### Bug fixes
 - When loading an MkDocs config, mike now runs the `startup` and `shutdown`